// src/shared/config/interfaces/IActionsTeams.ts
export interface IActionsTeam {
  BAD: {
    team: number;
    enemy: number;
  };
  ENTRY: {
    team: number;
    enemy: number;
  };
  FINE: {
    team: number;
    enemy: number;
  };
  GOAL: {
    team: number;
    enemy: number;
  };
  MISS: {
    team: number;
    enemy: number;
  };
  SWAP: {
    team: number;
    enemy: number;
  };
}